Change "children" prop to optional.
Refactor the Microsoft SSO login functionality to provide the user with a choice between the default login button and the custom component. The current version of the application enforces the use of the custom component for Microsoft SSO login, despite the presence of a default button. This update removes this limitation and enables the user to select their preferred login method.

By providing the user with the option to choose between the default button and the custom component, we increase the flexibility and usability of the application. Users who prefer the default login button can continue to use it, while those who require the additional functionality of the custom component can select that instead.

This change is expected to improve the overall user experience and make the application more accessible to a wider audience.